Revert build changes, document coverage fix plan
