+++ coverage_filtered.info
@@ -1,0 +1,304 @@
+TN:
+SF:/app/wrapper.c
+FN:168,179,decode_opj_common
+FN:181,187,get_alpha_component
+FN:189,221,write_headers_argb8888
+FN:223,266,write_headers_rgb565
+FN:25,27,getLastError
+FN:268,363,convert_image_to_bmp
+FN:35,43,opj_read_from_buffer
+FN:366,374,decodeToBmp
+FN:377,385,decodeToBmpWithRatio
+FN:388,430,getSize
+FN:45,54,get_codec_format
+FN:56,67,create_decoder
+FN:69,75,create_mem_stream
+FN:77,166,decode_internal
+FNDA:44,decode_opj_common
+FNDA:9,get_alpha_component
+FNDA:7,write_headers_argb8888
+FNDA:3,write_headers_rgb565
+FNDA:1,getLastError
+FNDA:13,convert_image_to_bmp
+FNDA:3,opj_read_from_buffer
+FNDA:38,decodeToBmp
+FNDA:6,decodeToBmpWithRatio
+FNDA:6,getSize
+FNDA:43,get_codec_format
+FNDA:43,create_decoder
+FNDA:40,create_mem_stream
+FNDA:39,decode_internal
+FNF:14
+FNH:14
+DA:25,1
+DA:26,1
+DA:35,3
+DA:36,3
+DA:37,3
+DA:38,3
+DA:39,2
+DA:40,2
+DA:41,2
+DA:42,2
+DA:45,43
+DA:46,43
+DA:47,43
+DA:48,43
+DA:49,43
+DA:50,43
+DA:51,1
+DA:53,42
+DA:56,43
+DA:57,43
+DA:58,43
+DA:61,41
+DA:62,41
+DA:63,1
+DA:64,1
+DA:66,40
+DA:69,40
+DA:70,40
+DA:71,40
+DA:72,40
+DA:73,40
+DA:74,40
+DA:77,39
+DA:78,39
+DA:80,39
+DA:82,39
+DA:83,39
+DA:84,2
+DA:85,2
+DA:88,37
+DA:90,37
+DA:91,37
+DA:92,3
+DA:94,34
+DA:95,34
+DA:98,34
+DA:99,6
+DA:100,6
+DA:101,6
+DA:102,6
+DA:103,6
+DA:104,6
+DA:106,28
+DA:107,28
+DA:108,28
+DA:109,28
+DA:125,34
+DA:126,34
+DA:128,34
+DA:129,29
+DA:130,17
+DA:132,12
+DA:134,1
+DA:139,34
+DA:140,18
+DA:141,18
+DA:142,18
+DA:143,18
+DA:144,2
+DA:145,2
+DA:147,2
+DA:148,1
+DA:149,1
+DA:150,1
+DA:151,1
+DA:152,1
+DA:153,1
+DA:154,1
+DA:156,14
+DA:157,10
+DA:158,10
+DA:159,10
+DA:162,37
+DA:163,37
+DA:165,37
+DA:168,44
+DA:169,44
+DA:170,44
+DA:172,44
+DA:173,5
+DA:174,5
+DA:177,39
+DA:178,39
+DA:181,9
+DA:182,9
+DA:183,30
+DA:184,25
+DA:186,5
+DA:189,7
+DA:190,7
+DA:191,7
+DA:192,7
+DA:195,7
+DA:196,7
+DA:197,7
+DA:198,7
+DA:199,7
+DA:200,7
+DA:203,7
+DA:204,7
+DA:205,7
+DA:206,7
+DA:207,7
+DA:208,7
+DA:209,7
+DA:210,7
+DA:211,7
+DA:212,7
+DA:213,7
+DA:214,7
+DA:215,7
+DA:216,7
+DA:217,7
+DA:218,7
+DA:219,7
+DA:220,7
+DA:221,7
+DA:223,3
+DA:224,3
+DA:225,3
+DA:226,3
+DA:227,3
+DA:230,3
+DA:231,3
+DA:232,3
+DA:233,3
+DA:234,3
+DA:235,3
+DA:238,3
+DA:239,3
+DA:240,3
+DA:241,3
+DA:242,3
+DA:243,3
+DA:244,3
+DA:245,3
+DA:246,3
+DA:247,3
+DA:248,3
+DA:249,3
+DA:250,3
+DA:251,3
+DA:252,3
+DA:253,3
+DA:254,3
+DA:255,3
+DA:260,3
+DA:261,3
+DA:262,3
+DA:263,3
+DA:264,3
+DA:265,3
+DA:266,3
+DA:268,13
+DA:269,13
+DA:270,13
+DA:272,13
+DA:273,1
+DA:274,1
+DA:277,12
+DA:278,12
+DA:279,12
+DA:280,12
+DA:282,12
+DA:283,2
+DA:284,2
+DA:285,2
+DA:286,10
+DA:287,1
+DA:288,1
+DA:289,1
+DA:290,1
+DA:291,1
+DA:294,9
+DA:295,9
+DA:296,9
+DA:297,9
+DA:300,12
+DA:302,12
+DA:304,4
+DA:305,4
+DA:306,4
+DA:307,4
+DA:309,4
+DA:310,4
+DA:311,1
+DA:312,1
+DA:315,3
+DA:317,3
+DA:318,16
+DA:319,13
+DA:320,119
+DA:321,106
+DA:322,106
+DA:323,106
+DA:324,106
+DA:325,106
+DA:327,13
+DA:332,8
+DA:333,8
+DA:334,8
+DA:335,8
+DA:337,8
+DA:338,8
+DA:339,1
+DA:340,1
+DA:343,7
+DA:345,7
+DA:346,7
+DA:347,20112
+DA:348,20106
+DA:349,20106
+DA:350,20106
+DA:351,20106
+DA:354,3
+DA:355,2
+DA:356,2
+DA:357,2
+DA:358,2
+DA:362,10
+DA:366,38
+DA:367,38
+DA:368,38
+DA:370,2
+DA:372,2
+DA:373,2
+DA:377,6
+DA:378,6
+DA:379,6
+DA:381,2
+DA:383,2
+DA:384,2
+DA:388,6
+DA:389,6
+DA:390,6
+DA:391,2
+DA:392,2
+DA:395,4
+DA:397,4
+DA:399,4
+DA:400,4
+DA:401,1
+DA:402,1
+DA:405,3
+DA:407,3
+DA:408,3
+DA:410,3
+DA:411,1
+DA:413,2
+DA:414,2
+DA:416,2
+DA:417,2
+DA:418,1
+DA:419,1
+DA:421,1
+DA:423,2
+DA:426,3
+DA:427,3
+DA:429,3
+LF:269
+LH:269
+end_of_record
